@@ -10,7 +10,7 @@ class ExampleTest extends TestCase
      *
      * @return void
      */
-    public function testBasicTest()
+    public function testBasicTest(): void
     {
         $this->assertTrue(true);
     }
